rename "admin" to dispatchadmin
