+++ dispatch/urls.py
@@ -1,3 +1,3 @@
-from dispatch.admin import urls as admin_urls
+from dispatch.dispatchadmin import urls as admin_urls
 from dispatch.api import urls as api_urls
 from dispatch.modules.podcasts import urls as podcasts_urls
